requirements.txt
Pillow==10.1.0
Pandas==2.1.4
streamlit==1.30.0
DateTime==5.4